fix: mudei a função de baixar ordem de compra PDF para outro arquivo, consertei o endpoint de mudar quantidade atual no histórico de ordem de compra junto com as partes que faltavam da tela
